refact(Statistic): rename
