+++ src/components/layouts/main-menu/main-menu.tsx
@@ -7,7 +7,7 @@ import Balances from './balances';
 import Invoices from './invoices';
 import OrganizationItem from './organization-item';
 import ProfileItem from './profile-item';
-import Statistics from './statistics';
+import Statistic from './statistic';
 
 const MainMenu = () => (
   <Navbar inverse fluid>
@@ -24,7 +24,7 @@ const MainMenu = () => (
       </Nav>
 
       <Nav pullRight>
-        <Statistics />
+        <Statistic />
         <Invoices />
         <ProfileItem />
       </Nav>
