+++ src/components/layouts/main-menu/statistic.tsx
@@ -3,12 +3,12 @@ import * as React from 'react';
 import { NavItem } from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
 
-import { statisticsPath } from 'routes';
+import { statisticPath } from 'routes';
 
 const Statistics: React.SFC = () => (
-  <LinkContainer to={ statisticsPath() }>
+  <LinkContainer to={ statisticPath() }>
     <NavItem >
-      Statistics
+      Statistic
     </NavItem>
   </LinkContainer>
 );
