+++ src/components/statistic/balances-chart.tsx
