+++ src/components/statistic/index.ts
